+++ Android/app/src/main/AndroidManifest.xml
@@ -9,24 +9,24 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/Theme.AndroidGUI">
-        <activity android:name="SelectEditOrDeleteActivity"
+        <activity android:name=".ManagerActivities.SelectEditOrDeleteActivity"
                   android:exported="true">
         </activity>
-        <activity android:name="AddStaffActivity">
+        <activity android:name=".ManagerActivities.AddStaffActivity">
         </activity>
         <activity
-                android:name="SeeReviewActivity"
+                android:name=".ManagerActivities.SeeReviewActivity"
                 android:exported="true">
         </activity>
-        <activity android:name="DeleteReviewActivity"
+        <activity android:name=".ManagerActivities.DeleteReviewActivity"
                   android:exported="true">
         </activity>
         <activity
-                android:name="ManagerPickActionActivity"
+                android:name=".ManagerActivities.ManagerPickActionActivity"
                 android:exported="true">
         </activity>
         <activity
-                android:name="ManageMenuActivity"
+                android:name=".ManagerActivities.ManageMenuActivity"
                 android:exported="true">
         </activity>
         <activity
