+++ Android/app/src/main/java/com/example/androidgui/ManagerActivities/DeleteReviewActivity.java
@@ -1,4 +1,4 @@
-package com.example.androidgui;
+package com.example.androidgui.ManagerActivities;
 
 import android.content.Intent;
 import android.os.Build;
@@ -8,6 +8,7 @@ import android.widget.NumberPicker;
 import android.widget.TextView;
 import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
+import com.example.androidgui.R;
 import constant.mangerSystem.ManagerDecision;
 import constant.uiMessage.ManagerUIMessage;
 import controller.reviewSystem.ReviewController;
