+++ Android/app/src/main/res/layout/activity_add_staff.xml
@@ -5,6 +5,6 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="AddStaffActivity">
+        tools:context=".ManagerActivities.AddStaffActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
