+++ Android/app/src/main/res/layout/activity_manager_pick_action.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="ManagerPickActionActivity">
+        tools:context=".ManagerActivities.ManagerPickActionActivity">
 
 
     <TextView
